Move repositories folder to src root
