+++ src/repositories/video/dto/trending-chain.dto.ts
@@ -1,4 +1,4 @@
-import { createProjectionFromDto } from "../../../../utils/createProjectionFromDto";
+import { createProjectionFromDto } from "../../../utils/createProjectionFromDto";
 
 export class TrendingChainDto {
   readonly created: Date;
